+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fixed incompatibility with Node 18/20 when using `--proxy` argument.
 - Append proxy configuration to `JAVA_OPTS` instead of replacing it.\
   This allows the caller to provide other Java options to the dependency-check-cli.
+- Validate GitHub release information before downloading the dependency-check-cli.
 
 ## Version 0.8.1